refactor: EmailInput, PasswordInput을 Input으로 합침
