+++ src/utils/globalStyle.js
@@ -1,6 +1,7 @@
-export const UIStyle = {
+const globalStyle = {
   inputStyle:
     'py-2 px-3 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-400 focus:border-transparent',
   buttonStyle:
     'border p-2 rounded-lg bg-blue-500 text-white hover:bg-blue-600 disabled:bg-blue-300',
 };
+export default globalStyle;
